- feat(protect): add user to the goals and make them private access

+++ controllers/goal.js
@@ -6,17 +6,18 @@ const Goal = require("../models/goal");
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
   const goals = await Goal.find();
-  res.status(200).json(goals);
+  res.status(201).json(goals);
 });
 // @desc   Create Goal
 // @route  POST /api/goals
 // @access Private
 const createGoal = asyncHandler(async (req, res) => {
   if (!req.body.text) {
-    res.status(400);
-    throw new Error("Please add a new field");
+    res.status(401);
+    throw new Error("Please add a text field");
   }
   const goal = await Goal.create({
+    user: req.user.id, // user form the middleware
     text: req.body.text,
   });
   res.status(201).json(goal);
@@ -29,12 +30,23 @@ const deleteGoal = asyncHandler(async (req, res) => {
   const goalID = req.params.id;
   const goal = await Goal.findById(goalID);
   if (!goal) {
-    res.status(400);
+    res.status(401);
     throw new Error("Goal not found");
+  }
+  // Check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
   await goal.deleteOne();
   //  await Goal.findByIdAndDelete(goalID);
-  res.status(200).json({ id: goalID });
+  res.status(201).json({ id: goalID });
 });
 
 // @desc   Update Goal
@@ -44,13 +56,24 @@ const updateGoal = asyncHandler(async (req, res) => {
   const goalID = req.params.id;
   const goal = await Goal.findById(goalID);
   if (!goal) {
-    res.status(400);
+    res.status(401);
     throw new Error("Goal not found");
+  }
+  // Check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
   const updatedGoal = await Goal.findByIdAndUpdate(goalID, req.body, {
     new: true,
   });
-  res.status(200).json(updatedGoal);
+  res.status(201).json(updatedGoal);
 });
 
 module.exports = {
